Refactor: extract macro runtime classes (Captor, DefaultValueProvider)
